refactor variable selector utils
